requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==4.2.5
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-filter==23.2
django-js-asset==2.1.0
django-location-field==2.7.2
django-phonenumber-field==7.1.0
django-solo==2.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-spectacular==0.26.4
factory-boy==3.3.0
Faker==19.6.1
fixtures==4.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
oauthlib==3.2.2
packaging==23.1
pbr==5.11.1
phonenumbers==8.13.20
Pillow==10.0.0
pluggy==1.3.0
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.2
pytest-django==4.5.2
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.2
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
testtools==2.6.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.5
